split up the last of render.py
